fix: ts formatting and linting

+++ packages/phspectra/tests/test_persistence.py
@@ -15,7 +15,7 @@ def _make_gaussian(x: npt.NDArray[np.float64], amp: float, mu: float, sig: float
 
 def test_single_peak() -> None:
     """A single Gaussian should produce exactly one persistent peak."""
-    x = np.linspace(0, 100, 500)
+    x = np.linspace(0, 100, 500, dtype=np.float64)
     signal = _make_gaussian(x, amp=5.0, mu=50.0, sig=3.0)
     peaks = find_peaks_by_persistence(signal)
     assert len(peaks) == 1
@@ -24,7 +24,7 @@ def test_single_peak() -> None:
 
 def test_two_peaks() -> None:
     """Two well-separated Gaussians should produce two significant peaks."""
-    x = np.linspace(0, 100, 500)
+    x = np.linspace(0, 100, 500, dtype=np.float64)
     signal = _make_gaussian(x, 5.0, 30.0, 3.0) + _make_gaussian(x, 3.0, 70.0, 4.0)
     peaks = find_peaks_by_persistence(signal, min_persistence=1.0)
     assert len(peaks) == 2
@@ -51,7 +51,7 @@ def test_plateau_same_component() -> None:
 
 def test_fit_gaussians_recovers_params() -> None:
     """fit_gaussians should approximately recover ground-truth parameters."""
-    x = np.linspace(0, 100, 500)
+    x = np.linspace(0, 100, 500, dtype=np.float64)
     signal = _make_gaussian(x, 5.0, 50.0, 5.0)
     components = fit_gaussians(signal)
     assert len(components) == 1
